requirements.txt
python-dotenv
fastapi==0.103.1
uvicorn==0.23.2
requests==2.31.0